+++ src/py_gql/execution/default_resolver.py
@@ -12,6 +12,7 @@ def default_resolver(
     root: Any,
     context: Any,
     info: "ResolveInfo",
+    *,
     __isinstance: Any = isinstance,
     __getattr: Any = getattr,
     __callable: Any = callable,